REFACTOR: Decomposed utils.ts
