+++ components/Account/AvatarSection.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import useLoadImage from "@/hooks/useLoadImage";
-import { getInitials } from "@/lib/utils";
+import { getInitials } from "@/lib/avatar/get-initials";
 import uploadUserAvatar from "@/actions/user/upload-user-avatar";
 import deleteUserAvatar from "@/actions/user/delete-user-avatar";
 import { AvatarFileSchema } from "@/schemas/user/avatar-file.schema";
