+++ components/Album/AlbumItem.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 
 import type { AlbumWithArtists } from "../../types/album-with-artists";
 import useLoadImage from "@/hooks/useLoadImage";
-import { formatArtists } from "@/lib/utils";
+import { formatArtists } from "@/lib/music/format-artists";
 import {
   Item,
   ItemContent,
