+++ components/Album/AlbumsGrid.tsx
@@ -4,7 +4,8 @@ import { useRouter } from "next/navigation";
 
 import type { AlbumWithArtists } from "../../types/album-with-artists";
 import { ROUTES } from "@/routes";
-import { cn, GRID_CLASSES } from "@/lib/utils";
+import { cn } from "@/lib/utils";
+import { GRID_CLASSES } from "@/lib/grid-classes";
 import AlbumItem from "@/components/Album/AlbumItem";
 
 interface AlbumsGridProps {
