+++ components/Player/PlayerContent.tsx
@@ -6,7 +6,8 @@ import { ListPlus, Info, ListMusic } from "lucide-react";
 import usePlayer from "@/hooks/usePlayer";
 import useLoadImage from "@/hooks/useLoadImage";
 import type { SongWithAlbum } from "../../types/song-with-album";
-import { cn, formatArtists } from "@/lib/utils";
+import { cn } from "@/lib/utils";
+import { formatArtists } from "@/lib/music/format-artists";
 import PlayerControls from "./PlayerControls";
 import PlayerVolume from "./PlayerVolume";
 import PlayerScrubber from "./PlayerScrubber";
