+++ components/Player/QueueSongItem.tsx
@@ -2,7 +2,8 @@
 
 import Image from "next/image";
 import { GripVertical, X } from "lucide-react";
-import { cn, formatArtists } from "@/lib/utils";
+import { cn } from "@/lib/utils";
+import { formatArtists } from "@/lib/music/format-artists";
 import useLoadImage from "@/hooks/useLoadImage";
 import type { SongWithAlbum } from "../../types/song-with-album";
 /**
