+++ components/Song/SongOptionsMenu.tsx
@@ -10,7 +10,8 @@ import useFavourite from "@/hooks/useFavourite";
 import useAddToPlaylist from "@/hooks/useAddToPlaylist";
 import { useUser } from "@/hooks/useUser";
 import usePlayer from "@/hooks/usePlayer";
-import { cn, formatArtists } from "@/lib/utils";
+import { cn } from "@/lib/utils";
+import { formatArtists } from "@/lib/music/format-artists";
 import deleteSong from "@/actions/song/delete-song";
 import type { SongWithAlbum } from "../../types/song-with-album";
 import { Button } from "@/components/ui/button";
