+++ lib/music/format-artists.ts
@@ -1,0 +1,20 @@
+import type { AlbumWithArtists } from "@/types/album-with-artists";
+
+/**
+ * Formats an album's artist list into a human-readable string.
+ * Used for displaying artist attribution in song/album cards and detail pages.
+ *
+ * @param album - Album object with artist relationships
+ * @returns Comma-separated artist names, or "Unknown Artist" if empty
+ *
+ * @example
+ * formatArtists({ ...album, artists: [{ name: "John" }, { name: "Jane" }] })
+ * // Output: "John, Jane"
+ *
+ * formatArtists({ ...album, artists: [] })
+ * // Output: "Unknown Artist"
+ */
+export function formatArtists(album: AlbumWithArtists): string {
+  if (!album.artists.length) return "Unknown Artist";
+  return album.artists.map((a) => a.name).join(", ");
+}
